Feature #70119, refactor out VSAssemblyDndController to set up service proxy for children classes
