+++ core/src/main/java/inetsoft/web/binding/dnd/controller/VSCalcTableDndController.java
@@ -17,44 +17,25 @@
  */
 package inetsoft.web.binding.dnd.controller;
 
-import inetsoft.analytic.composition.ViewsheetService;
-import inetsoft.report.composition.RuntimeViewsheet;
-import inetsoft.uql.asset.AggregateInfo;
 import inetsoft.uql.asset.SourceInfo;
-import inetsoft.uql.erm.DataRef;
-import inetsoft.uql.viewsheet.CalcTableVSAssembly;
 import inetsoft.uql.viewsheet.VSAssembly;
 import inetsoft.uql.viewsheet.internal.*;
-import inetsoft.web.binding.controller.VSTableLayoutController;
-import inetsoft.web.binding.controller.VSTableLayoutService;
-import inetsoft.web.binding.dnd.CalcDropTarget;
-import inetsoft.web.binding.dnd.CalcTableTransfer;
 import inetsoft.web.binding.event.VSDndEvent;
-import inetsoft.web.binding.handler.VSAssemblyInfoHandler;
-import inetsoft.web.binding.handler.VSCalcTableBindingHandler;
-import inetsoft.web.binding.service.VSBindingService;
 import inetsoft.web.viewsheet.LoadingMask;
 import inetsoft.web.viewsheet.Undoable;
 import inetsoft.web.viewsheet.model.RuntimeViewsheetRef;
-import inetsoft.web.viewsheet.model.VSObjectModelFactoryService;
 import inetsoft.web.viewsheet.service.*;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.messaging.handler.annotation.MessageMapping;
 import org.springframework.messaging.handler.annotation.Payload;
 import org.springframework.stereotype.Controller;
-
-import java.awt.*;
 import java.security.Principal;
-import java.util.List;
-import java.util.Set;
 
 /**
  * This class handles get vsobjectmodel from the server.
  */
 @Controller
-public class VSCalcTableDndController extends VSAssemblyDndController {
+public class VSCalcTableDndController {
    /**
     * Creates a new instance of <tt>VSViewController</tt>.
     *
@@ -64,18 +45,11 @@ public class VSCalcTableDndController extends VSAssemblyDndController {
    @SuppressWarnings("SpringJavaInjectionPointsAutowiringInspection")
    @Autowired
    public VSCalcTableDndController(RuntimeViewsheetRef runtimeViewsheetRef,
-                                   VSBindingService bfactory,
-                                   VSAssemblyInfoHandler assemblyInfoHandler,
-                                   VSTableLayoutService tableLayoutService,
-                                   VSCalcTableBindingHandler calcTableHandler,
-                                   VSObjectModelFactoryService objectModelService,
-                                   ViewsheetService viewsheetService,
-                                   CoreLifecycleService coreLifecycleService)
+                                   VSCalcTableDndServiceProxy vsCalcTableDndServiceProxy)
    {
-      super(runtimeViewsheetRef, bfactory, assemblyInfoHandler, objectModelService,
-            viewsheetService, coreLifecycleService);
-      this.calcTableHandler = calcTableHandler;
-      this.tableLayoutService = tableLayoutService;
+
+      this.runtimeViewsheetRef = runtimeViewsheetRef;
+      this.vsCalcTableDndServiceProxy = vsCalcTableDndServiceProxy;
    }
 
    /**
@@ -88,20 +62,11 @@ public class VSCalcTableDndController extends VSAssemblyDndController {
    public void addRemoveColumns(@Payload VSDndEvent event, Principal principal,
       @LinkUri String linkUri, CommandDispatcher dispatcher) throws Exception
    {
-      RuntimeViewsheet rvs = getRuntimeVS(principal);
+      String runtimeId = runtimeViewsheetRef.getRuntimeId();
 
-      if(rvs == null) {
-         return;
+      if(runtimeId != null) {
+         vsCalcTableDndServiceProxy.addRemoveColumns(runtimeId, event, principal, linkUri, dispatcher);
       }
-
-      CalcTableVSAssembly assembly = (CalcTableVSAssembly)
-         getVSAssembly(rvs, event.name());
-      CalcTableVSAssembly clone = (CalcTableVSAssembly) assembly.clone();
-      CalcTableTransfer transfer = (CalcTableTransfer)event.getTransfer();
-      CalcDropTarget target = (CalcDropTarget)event.getDropTarget();
-      calcTableHandler.addRemoveColumns(clone, transfer.getDragRect(), target.getDropRect());
-      applyAssemblyInfo(rvs, assembly, (VSAssemblyInfo)clone.getInfo(), dispatcher,
-         event, linkUri, null);
    }
 
    /**
@@ -114,45 +79,13 @@ public class VSCalcTableDndController extends VSAssemblyDndController {
    public void addColumns(@Payload VSDndEvent event, Principal principal,
       @LinkUri String linkUri, CommandDispatcher dispatcher) throws Exception
    {
-      RuntimeViewsheet rvs = getRuntimeVS(principal);
-
-      if(rvs == null) {
-         return;
+      String runtimeId = runtimeViewsheetRef.getRuntimeId();
+      if(runtimeId != null) {
+         vsCalcTableDndServiceProxy.addColumns(runtimeId, event, principal,
+                                                 linkUri, dispatcher);
       }
-
-      CalcTableVSAssembly assembly =
-         (CalcTableVSAssembly) getVSAssembly(rvs, event.name());
-      CalcTableVSAssembly nassembly = (CalcTableVSAssembly) assembly.clone();
-      CalcTableVSAssemblyInfo ninfo = (CalcTableVSAssemblyInfo) nassembly.getInfo();
-
-      // Handle source changed.
-      if(sourceChanged(assembly, event.getTable())) {
-         changeSource(nassembly, event.getTable(), event.getSourceType());
-         CalcTableVSAssemblyInfo vsCalcTableInfo =
-            (CalcTableVSAssemblyInfo) nassembly.getVSAssemblyInfo();
-
-         if(vsCalcTableInfo != null) {
-            AggregateInfo ainfo =  vsCalcTableInfo.getAggregateInfo();
-
-            if(ainfo != null) {
-               List<DataRef> calcFields = ainfo.getFormulaFields();
-               Set<String> calcFieldsRefs = ainfo.removeFormulaFields(calcFields);
-               nassembly.getTableLayout().clearFormulaBinding(calcFieldsRefs);
-            }
-         }
-      }
-
-      if(ninfo.getSourceInfo() == null) {
-         ninfo.setSourceInfo(new SourceInfo(event.getSourceType(), null, event.getTable()));
-      }
-
-      CalcDropTarget target = (CalcDropTarget) event.getDropTarget();
-      calcTableHandler.addColumns(nassembly, event.getEntries(), target.getDropRect(), rvs);
-      applyAssemblyInfo(rvs, assembly, nassembly, dispatcher, event,
-         "/events/vscalctable/dnd/addColumns", linkUri);
    }
 
-   @Override
    protected boolean sourceChanged(VSAssembly assembly, String table) {
       SourceInfo sinfo = ((DataVSAssemblyInfo) assembly.getInfo()).getSourceInfo();
       return sinfo != null && !sinfo.getSource().equals(table);
@@ -168,42 +101,14 @@ public class VSCalcTableDndController extends VSAssemblyDndController {
    public void removeColumns(@Payload VSDndEvent event, Principal principal,
       @LinkUri String linkUri, CommandDispatcher dispatcher) throws Exception
    {
-      RuntimeViewsheet rvs = getRuntimeVS(principal);
-      if(rvs == null) {
-         return;
-      }
+      String runtimeId = runtimeViewsheetRef.getRuntimeId();
 
-      CalcTableVSAssembly assembly =
-         (CalcTableVSAssembly) getVSAssembly(rvs, event.name());
-      CalcTableVSAssembly clone = (CalcTableVSAssembly) assembly.clone();
-      CalcTableTransfer transfer = (CalcTableTransfer) event.getTransfer();
-      calcTableHandler.removeColumns(clone, transfer.getDragRect());
-      applyAssemblyInfo(rvs, assembly, (VSAssemblyInfo) clone.getInfo(), dispatcher, event,
-                        linkUri, null);
+      if(runtimeId != null) {
+         vsCalcTableDndServiceProxy.removeColumns(runtimeId, event,
+                                                  principal, linkUri, dispatcher);
+      }
    }
 
-   protected void createDndCommands(RuntimeViewsheet rvs, VSAssembly assembly,
-      CommandDispatcher dispatcher, VSDndEvent event, String linkUri) throws Exception
-   {
-      super.createDndCommands(rvs, assembly, dispatcher, event, linkUri);
-
-      String name = assembly.getInfo().getAbsoluteName();
-      Rectangle rect = null;
-
-      if(event.getDropTarget() != null) {
-         rect = ((CalcDropTarget) event.getDropTarget()).getDropRect();
-      }
-      else {
-         rect = ((CalcTableTransfer) event.getTransfer()).getDragRect();
-      }
-
-      CalcTableVSAssembly calc = (CalcTableVSAssembly) assembly;
-      dispatcher.sendCommand(name,
-         tableLayoutService.createCellBindingCommand(rvs, calc, rect.y, rect.x));
-      dispatcher.sendCommand(name, tableLayoutService.createTableLayoutCommand(rvs, calc));
-   }
-
-   private VSCalcTableBindingHandler calcTableHandler;
-   private VSTableLayoutService tableLayoutService;
-   private static final Logger LOG = LoggerFactory.getLogger(VSCalcTableDndController.class);
+   private RuntimeViewsheetRef runtimeViewsheetRef;
+   private VSCalcTableDndServiceProxy vsCalcTableDndServiceProxy;
 }
